Making the meeting ID an attribute of the scraper.
This changes the way a new scraper is instantiated as you now need to pass the ID as the first argument.

This also changes the behviour of the domain argument as this now has to be issued as keyword argument.

To seperate the concers of our objects there is now a new RISInformation base object that holds the has_website_changed-method. With DarmstadtRIS there exists a subclass of RISInformation that has a correct BASE_URL set. Maybe this is an approach we wan't to use for other objects aswell to make this scraper as versatile as possible without the need to pass the city to every object. This also could be used to implement scraping of parts that differ between citys.